Basic project outline.
Set up a Makefile for a single directory project, as well as .h and .cpp
files for block and gameboard.

Put in stubs for block.cpp as per the UML diagram and some failing tests
in tests.cpp which should pass when the constructor and first few
methods are complete.